requirements.txt
Django>=4.2.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0 
streamlit>=1.32.0
streamlit-folium>=0.15.0
folium>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
altair>=5.0.0
requests>=2.31.0 